# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: "http://127.0.0.1:41727" | (string & {});
};

export type HealthResponse = {
  status: "ok";
};

export type RepositoryLocator = {
  provider: "github";
  slug: string;
};

export type Project = {
  id: string;
  repository: RepositoryLocator;
};

export type ProjectsResponse = {
  projects: Array<Project>;
};

export type Milestone = {
  number: number;
  title: string;
  state: OpenClosedState;
};

export type MilestonesResponse = {
  projectId: string;
  milestones: Array<Milestone>;
};

export type Issue = {
  number: number;
  title: string;
  state: OpenClosedState;
  labels: Array<string>;
  dependency: DependencyJudgment;
};

export type MilestoneIssuesResponse = {
  projectId: string;
  milestoneNumber: number;
  issues: Array<Issue>;
};

export type IssueResponse = {
  projectId: string;
  issue: Issue;
};

export type OpenClosedState = "OPEN" | "CLOSED";

export type DependencyJudgment = DependencyReady | DependencyBlocked | DependencyError;

export type DependencyReady = {
  status: "READY";
};

export type DependencyBlocked = {
  status: "BLOCKED";
  blockedBy: Array<number>;
};

export type DependencyError = {
  status: "ERROR";
  error: ErrorDetail;
};

export type ErrorDetail = {
  code: string;
  message: string;
};

export type ErrorResponse = {
  error: ErrorDetail;
};

export type ProjectId = string;

export type MilestoneNumber = number;

export type IssueNumber = number;

export type GetHealthData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/healthz";
};

export type GetHealthResponses = {
  /**
   * Server is healthy
   */
  200: HealthResponse;
};

export type GetHealthResponse = GetHealthResponses[keyof GetHealthResponses];

export type ListProjectsData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/api/v1/projects";
};

export type ListProjectsResponses = {
  /**
   * Configured Projects
   */
  200: ProjectsResponse;
};

export type ListProjectsResponse = ListProjectsResponses[keyof ListProjectsResponses];

export type ListProjectMilestonesData = {
  body?: never;
  path: {
    projectId: string;
  };
  query?: never;
  url: "/api/v1/projects/{projectId}/milestones";
};

export type ListProjectMilestonesErrors = {
  /**
   * The Project is not configured
   */
  404: ErrorResponse;
  /**
   * Current GitHub facts could not be read completely
   */
  502: ErrorResponse;
};

export type ListProjectMilestonesError =
  ListProjectMilestonesErrors[keyof ListProjectMilestonesErrors];

export type ListProjectMilestonesResponses = {
  /**
   * Current Milestones
   */
  200: MilestonesResponse;
};

export type ListProjectMilestonesResponse =
  ListProjectMilestonesResponses[keyof ListProjectMilestonesResponses];

export type ListMilestoneIssuesData = {
  body?: never;
  path: {
    projectId: string;
    milestoneNumber: number;
  };
  query?: never;
  url: "/api/v1/projects/{projectId}/milestones/{milestoneNumber}/issues";
};

export type ListMilestoneIssuesErrors = {
  /**
   * A path number is invalid
   */
  400: ErrorResponse;
  /**
   * The Project is not configured
   */
  404: ErrorResponse;
  /**
   * Current GitHub facts could not be read completely
   */
  502: ErrorResponse;
};

export type ListMilestoneIssuesError = ListMilestoneIssuesErrors[keyof ListMilestoneIssuesErrors];

export type ListMilestoneIssuesResponses = {
  /**
   * Current Milestone Issues and dependency judgments
   */
  200: MilestoneIssuesResponse;
};

export type ListMilestoneIssuesResponse =
  ListMilestoneIssuesResponses[keyof ListMilestoneIssuesResponses];

export type GetProjectIssueData = {
  body?: never;
  path: {
    projectId: string;
    issueNumber: number;
  };
  query?: never;
  url: "/api/v1/projects/{projectId}/issues/{issueNumber}";
};

export type GetProjectIssueErrors = {
  /**
   * A path number is invalid
   */
  400: ErrorResponse;
  /**
   * The Project is not configured
   */
  404: ErrorResponse;
  /**
   * Current GitHub facts could not be read completely
   */
  502: ErrorResponse;
};

export type GetProjectIssueError = GetProjectIssueErrors[keyof GetProjectIssueErrors];

export type GetProjectIssueResponses = {
  /**
   * Current Issue and dependency judgment
   */
  200: IssueResponse;
};

export type GetProjectIssueResponse = GetProjectIssueResponses[keyof GetProjectIssueResponses];